requirements.txt
streamlit
transformers
torch
chromadb
cohere
deep-translator
gTTS
Pillow
nltk
pytesseract
PyMuPDF
newspaper3k
beautifulsoup4